Replace sqrt(atoi... with sqrt(atof... which was causing a compilation error with PGI on Yellowstone.
Former-commit-id: 3c5ee15d0bfddb78bb82fd3609c8d865a4242174

+++ met/src/tools/other/gsi_tools/gsi_util.cc
@@ -298,7 +298,7 @@ void setup_header(StatHdrColumns &shc,
 
    // INTERP_PNTS
    if(name.has("INTERP_PNTS", index)) {
-      shc.set_interp_wdth(nint(sqrt(atoi(value[index]))));
+      shc.set_interp_wdth(nint(sqrt(atof(value[index]))));
    }
    else {
       shc.set_interp_wdth(default_interp_wdth);
